Fix types in data_chunk.rs

+++ crates/messages/src/data_chunk.rs
@@ -2,14 +2,14 @@ use crate::Range;
 
 #[derive(Clone, Ord, PartialOrd, Eq, PartialEq)]
 pub struct DataChunk {
-    top: u32,
-    first_block: u32,
-    last_block: u32,
+    top: u64,
+    first_block: u64,
+    last_block: u64,
     last_hash: String,
 }
 
 impl DataChunk {
-    pub fn new(top: u32, first_block: u32, last_block: u32, last_hash: String) -> Self {
+    pub fn new(top: u64, first_block: u64, last_block: u64, last_hash: String) -> Self {
         assert!(top <= first_block);
         assert!(first_block <= last_block);
         Self {
@@ -21,17 +21,17 @@ impl DataChunk {
     }
 
     #[inline]
-    pub fn top(&self) -> u32 {
+    pub fn top(&self) -> u64 {
         self.top
     }
 
     #[inline]
-    pub fn first_block(&self) -> u32 {
+    pub fn first_block(&self) -> u64 {
         self.first_block
     }
 
     #[inline]
-    pub fn last_block(&self) -> u32 {
+    pub fn last_block(&self) -> u64 {
         self.last_block
     }
 }
@@ -58,11 +58,11 @@ impl std::str::FromStr for DataChunk {
     #[allow(clippy::get_first)]
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         let top_range_split = s.split('/').collect::<Vec<_>>();
-        let top: u32 = top_range_split.get(0).ok_or(())?.parse().or(Err(()))?;
+        let top: u64 = top_range_split.get(0).ok_or(())?.parse().or(Err(()))?;
         let range_str = top_range_split.get(1).ok_or(())?;
         let range_split = range_str.split('-').collect::<Vec<_>>();
-        let beg: u32 = range_split.get(0).ok_or(())?.parse().or(Err(()))?;
-        let end: u32 = range_split.get(1).ok_or(())?.parse().or(Err(()))?;
+        let beg: u64 = range_split.get(0).ok_or(())?.parse().or(Err(()))?;
+        let end: u64 = range_split.get(1).ok_or(())?.parse().or(Err(()))?;
         let hash = range_split.get(2).ok_or(())?;
         if top <= beg && beg <= end {
             Ok(Self::new(top, beg, end, (*hash).to_string()))
